--- web/src/redcross/discussionModule/components/Discussion.less
@@ -1,24 +1,0 @@
-@button-background: #dddddd;
-
-.content {
-  margin-left: auto;
-  margin-right: auto;
-  width: 70%;
-}
-
-.container {
-  padding-bottom: 70px;
-}
-
-.title {
-  color: red;
-}
-
-.add-button {
-  background-color: @button-background;
-  float: right;
-}
-
-textarea {
-  resize: none;
-}
