+++ web/src/redcross/discussionModule/components/DiscussionContainer.less
@@ -1,0 +1,34 @@
+@button-background: #dddddd;
+
+.container {
+
+  padding-bottom: 70px;
+
+  .content {
+
+    margin-left: auto;
+    margin-right: auto;
+    width: 70%;
+
+    .title {
+
+      color: red;
+
+    }
+
+    .add-button {
+
+      background-color: @button-background;
+      float: right;
+
+    }
+
+  }
+
+}
+
+textarea {
+
+  resize: none;
+
+}
