+++ web/src/redcross/discussionModule/components/DiscussionContainer.ng.html
