+++ web/src/redcross/discussionModule/components/DiscussionPost.less
@@ -2,16 +2,22 @@
 @dark-gray: #bbbbbb;
 
 .container {
+
   margin-top: 20px;
   margin-bottom: 20px;
-}
 
-.content-container {
-  background-color: @light-gray;
-  border: 1px solid @dark-gray;
-  border-radius: 5px;
-}
+  .content-container {
 
-.content {
-  margin: 5px;
+    background-color: @light-gray;
+    border: 1px solid @dark-gray;
+    border-radius: 5px;
+
+    .content {
+
+      margin: 5px;
+
+    }
+
+  }
+
 }
